Added basic decimal integer to string function

+++ src/types.h
@@ -1,6 +1,6 @@
 #pragma once
 
-typedef unsigned long int   u64;
+typedef unsigned long long  u64;
 typedef unsigned int        u32;
 typedef unsigned short int  u16;
 typedef unsigned char       u8;
